requirements.txt
yt-dlp
discord.py
pynacl